backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.20.0",
    "alembic>=1.14.1",
    "asyncpg>=0.30.0",
    "boto3>=1.35.0",
    "fastapi>=0.124.4",
    "greenlet>=3.0.0",
    "httpx>=0.28.1",
    "numpy>=1.24.4",
    "opencv-python>=4.13.0.92",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.7.1",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "redis>=6.1.1",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.33.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.15.7",
]
